Refactor packages and remove sample structures
