+++ bundles/org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface/META-INF/MANIFEST.MF
@@ -15,4 +15,3 @@ Require-Bundle: javax.inject,
 Bundle-RequiredExecutionEnvironment: JavaSE-11
 Import-Package: javax.annotation;version="1.2.0"
 Automatic-Module-Name: org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface
-Export-Package: org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface.parts.interfaces
