+++ bundles/org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis/META-INF/MANIFEST.MF
@@ -11,7 +11,6 @@ Require-Bundle: javax.inject,
  org.eclipse.e4.ui.services,
  org.eclipse.e4.core.di.annotations,
  org.palladiosimulator.pcm,
- org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface,
  org.jsoup;bundle-version="1.8.3",
  org.palladiosimulator.pcm.confidentiality.attacker,
  org.apache.log4j,
@@ -19,4 +18,5 @@ Require-Bundle: javax.inject,
 Bundle-RequiredExecutionEnvironment: JavaSE-11
 Automatic-Module-Name: org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis
 Import-Package: javax.annotation
-Export-Package: org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.parts
+Export-Package: org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis,
+ org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface
