+++ bundles/org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis/src/org/palladiosimulator/pcm/confidentiality/reverseengineering/staticcodeanalysis/NvdVulnerabilityDatabase.java
@@ -1,4 +1,4 @@
-package org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.parts;
+package org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis;
 
 import java.io.IOException;
 
@@ -6,8 +6,8 @@ import org.jsoup.Jsoup;
 import org.palladiosimulator.pcm.confidentiality.attackerSpecification.attackSpecification.CVEVulnerability;
 import org.palladiosimulator.pcm.confidentiality.attackerSpecification.attackSpecification.CWEVulnerability;
 import org.palladiosimulator.pcm.confidentiality.attackerSpecification.attackSpecification.impl.AttackSpecificationFactoryImpl;
-import org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface.parts.interfaces.IVulnerabilityDatabase;
-import org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface.parts.interfaces.VulnerabilityDatabaseException;
+import org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface.IVulnerabilityDatabase;
+import org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface.VulnerabilityDatabaseException;
 import org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.nvd_rest_api.NvdResponse;
 
 import com.google.gson.Gson;
