+++ bundles/org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis/src/org/palladiosimulator/pcm/confidentiality/reverseengineering/staticcodeanalysis/SnykIssue.java
@@ -1,9 +1,9 @@
-package org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.parts;
+package org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis;
 
 import java.util.Date;
 import java.util.Map;
 
-import org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface.parts.interfaces.IStaticCodeAnalysisIssue;
+import org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface.IStaticCodeAnalysisIssue;
 
 public class SnykIssue implements IStaticCodeAnalysisIssue {
