+++ bundles/org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis/src/org/palladiosimulator/pcm/confidentiality/reverseengineering/staticcodeanalysis/StaticCodeAnalyisResult.java
@@ -1,9 +1,9 @@
-package org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.parts;
+package org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis;
 
 import java.util.List;
 
-import org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface.parts.interfaces.IStaticCodeAnalysisIssue;
-import org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface.parts.interfaces.IStaticCodeAnalysisResult;
+import org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface.IStaticCodeAnalysisIssue;
+import org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface.IStaticCodeAnalysisResult;
 
 public class StaticCodeAnalyisResult implements IStaticCodeAnalysisResult {
