+++ bundles/org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis/src/org/palladiosimulator/pcm/confidentiality/reverseengineering/staticcodeanalysis/iface/ICodeAnalysisResult.java
@@ -1,4 +1,4 @@
-package org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface.parts.interfaces;
+package org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface;
 
 /**
  * Empty parent interface for CodeAnalysResults. Enables more analysis methods
