+++ bundles/org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis/src/org/palladiosimulator/pcm/confidentiality/reverseengineering/staticcodeanalysis/iface/ICodeAnalyst.java
@@ -1,4 +1,4 @@
-package org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface.parts.interfaces;
+package org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface;
 
 /**
  * Parent interface for CodeAnalysts. Use this to enable new CodeAnalysts later
