+++ bundles/org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis/src/org/palladiosimulator/pcm/confidentiality/reverseengineering/staticcodeanalysis/iface/IStaticCodeAnalysisResult.java
@@ -1,4 +1,4 @@
-package org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface.parts.interfaces;
+package org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface;
 
 import java.util.List;
