+++ bundles/org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis/src/org/palladiosimulator/pcm/confidentiality/reverseengineering/staticcodeanalysis/iface/IVulnerabilityDatabase.java
@@ -1,4 +1,4 @@
-package org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface.parts.interfaces;
+package org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface;
 
 import org.palladiosimulator.pcm.confidentiality.attackerSpecification.attackSpecification.CVEVulnerability;
 import org.palladiosimulator.pcm.confidentiality.attackerSpecification.attackSpecification.CWEVulnerability;
