+++ bundles/org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis/src/org/palladiosimulator/pcm/confidentiality/reverseengineering/staticcodeanalysis/iface/VulnerabilityDatabaseException.java
@@ -1,9 +1,9 @@
-package org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface.parts.interfaces;
+package org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface;
 
 /**
  * Is to be thrown if access to a IVulnerabilityDatabase failed.
  * 
- * @see org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface.parts.interfaces.IVulnerabilityDatabase
+ * @see org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface.IVulnerabilityDatabase
  * 
  * @author Florian Bossert
  */
