+++ tests/org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.test/META-INF/MANIFEST.MF
@@ -12,7 +12,6 @@ Require-Bundle: javax.inject,
  org.eclipse.e4.core.di.annotations,
  org.eclipse.core.runtime,
  org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis,
- org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface,
  org.junit;bundle-version="4.13.0",
  org.jsoup;bundle-version="1.8.3",
  org.palladiosimulator.pcm,
