+++ tests/org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.test/src/org/palladiosimulator/pcm/confidentiality/reverseengineering/staticcodeanalysis/test/SnykParsingTest.java
@@ -5,8 +5,8 @@ import java.nio.file.Path;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.parts.SnykCLIStaticCodeAnalyst;
-import org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.parts.StaticCodeAnalyisResult;
+import org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.SnykCLIStaticCodeAnalyst;
+import org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.StaticCodeAnalyisResult;
 
 public class SnykParsingTest {
