+++ tests/org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.test/src/org/palladiosimulator/pcm/confidentiality/reverseengineering/staticcodeanalysis/test/SnykWebsiteParsingTest.java
@@ -11,8 +11,8 @@ import java.util.stream.Collectors;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 import org.junit.Test;
-import org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface.parts.interfaces.VulnerabilityDatabaseException;
-import org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.parts.NvdVulnerabilityDatabase;
+import org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.NvdVulnerabilityDatabase;
+import org.palladiosimulator.pcm.confidentiality.reverseengineering.staticcodeanalysis.iface.VulnerabilityDatabaseException;
 
 public class SnykWebsiteParsingTest {
     /**
